iota:  Add replacements for ctype.h functions
The functions and macros defined in <ctype.h> and <cctype> are dependent
on locale.  While a program is free to select a locale with setlocale(),
assuring expected behavior for non-ASCII code points, library code isn't
able to do this (without the risk of interfering with other parts of the
program).  The only viable option is to avoid locale-dependent functions
in library code which wants exactly '0' - '9' and not just any character
that could be construed as a digit, for example.  (Note that MSL doesn't
have any discrepancies regarding isdigit() in its default locale, but it
*does* return true from isalpha() for letters with diacritical marks (in
MacRoman).)

The new functions have different names, both to avoid issues with macros
and to prevent accidents in code that imports symbols from the namespace
iota instead of explicitly scoping.