gear:  Replace locale-dependent <cctype> functions
These functions are used for parsing, so they should avoid being locale-
dependent.

isdigit:  While isdigit() is unaffected by the current locale in ANSI C,
Microsoft's implementation returns non-zero for superscript digits using
code page 1252.  The string "\xB9\xB2\xB3" ("123", superscript) would be
parsed as a 3-digit integer with the value 15131.  If the caller were to
use the resulting value as an array subscript (under the assumption that
the input string length reliably indicated an upper limit on the value),
then it's conceivable that an information leak or stack corruption would
be possible.

isspace:  Metrowerks' implementation of isspace() in MSL 6 considers the
nonbreaking space (0xCA in MacRoman) to be whitespace.  (There can be no
distinction between horizontal and other whitespace, because MSL doesn't
provide isblank().)  At present, parse_float() isn't used in more than a
few places: to parse command-line arguments to be passed to nanosleep(),
and parsing 3D model data.  In the former case, input strings are solely
provided through argv (and presumably not passed along from an untrusted
source).  At worst, an argument is prefixed with 0xCA, which in MacRelix
is ignored and elsewhere halts parsing, resulting in a zero result or an
error -- so the same input might work as expected in one context and DoS
the CPU in another.  In the latter case, untrusted input is expected, as
this is a data document.  Again, it's possible that isspace() might have
different results on different platforms.

<http://stackoverflow.com/questions/2898228/can-isdigit-legitimately-be-locale-dependent-in-c>