[NEW] mixerfs{,d}:  A vfs interface to mixctl
The mixerfs library implements a vfs node with amixer properties such as
"volume" (and "route" on the Raspberry Pi).  To handle requests, it runs
mixctl in a coprocess (which in turn runs amixer).

The mixerfsd project is a Freemount server with the mixerfs directory as
root.  It communicates with a client on std/stdout (as freemountd does).