plus:  Remove unused "basic" buffer deletion mode
The idea of allocating an uninitialized buffer which the caller fills in
without having to zero-initialize it first is sound.  But plus::string's
reset() member is a superior interface for doing so.  Besides being safe
in the event of an exception, the call itself is easier to write without
looking it up and the resulting buffer can be shared.  The only downside
is that a default construction happens needlessly.