cpres:  Use Unicode filing routines when available
This changes the behavior of the --data option:  Previously, if a source
file's data fork couldn't be opened as a resource file (for any reason),
cpres would fall back to the resource fork.  Now it doesn't.  (A-line is
unaffected since it already Rezzes into the data fork for Mach-O, so the
source and destination for cpres use the same fork.)