plus:  Fix multiple extent_alloc() issues
We were setting header->capacity to the total extent size (including the
header).  Set it to just the storage capacity.

Disallowing a capacity with the high bit set is a problem because a test
tries to allocate a 0x7FFFFFFF-byte string buffer, which gets rounded up
to 0x80000000 to include the trailing NUL.  Allow any capacity such that
overflow doesn't occur when adding the extent header size.

Don't just assert.  It's likely that a caller will write attacker-chosen
data to the returned address.  If overflow occurs, call abort(), even in
non-debug contexts.

Security analysis:  The only extent_alloc() user so far is plus::string,
which has a max_size() of 0x7FFFFFFF[FFFFFFFF].  Provided its checks are
complete, it could not have requested more than 0x80000000[00000000], so
in the absence of any missed size checks, extent_alloc() wasn't at risk.