math:  Fix sum_size_desc() for non-equal size args
While we were careful to swap operands in ascending order, and we had an
additional loop to propagate the carry bit when add()'s second parameter
was shorter than the first, the sum_size routines didn't actually handle
arguments of unequal size.  Loop over the portion of the first parameter
where the second would have a leading zero, and exit on a limb not equal
to limb_t( -1 ).