A-line:  Check {CWANSI,CWC,C}Includes variables
Prefer directory locations provided by environment variables rather than
subdirectories of Interfaces&Libraries/Interfaces.

It's necessary to specify the search directory even when it isn't set by
an environment variable, so as not to affect the search order -- even if
the environment provides CIncludes but not CWANSIIncludes, we still have
to search the latter one first.