plus:  Add integer division and remainder
The percent operator is defined as remainder, not modulo.  A division by
zero throws an exception.

The division algorithm is recursive for efficiency when the dividend has
far greater magnitude than the divisor.  However, the operation fails if
there's insufficient stack space.  On modern operating systems (and also
in MacRelix), the process will be sent an appropriate signal (SIGSEGV or
SIGSTKFLT) in this case.  In environments without protected memory, it's
quite possible (and perhaps even likely) that a stack overrun will crash
(or otherwise corrupt) the entire system -- caveat executor.