vc:  Rewrite operator-out-of-context error message
This is not necessarily an internal error.  This code path executes when
a token is recognized as valid (and assumed to be an operator), but it's
not defined as an operator in that context.  For example, left unary `*`
would trigger this diagnostic.  (Then again, the same error still occurs
if the developer adds a new token and forgets to map it to an operator.)