vc:  Permit functions of boolean values
Yes, this means you can call abs() and half() with boolean values -- for
the time being (since this isn't supported and will probably go away) --
though it's useless to do so, because abs(flag) is the identity function
and half(flag) is always false.  The rationale here is to let bool(flag)
work (as an identity function).