vc:  Distinguish nothing / dummy operand / undefined
Nothing is a non-value state resulting from operations that don't return
values, such as simple assignment and the empty statement.

Dummy-operand, another non-value state, is a shim for filling the unused
operand slot in unary operations.

Undefined is the initial non-value state of a symbol (i.e. a constant or
variable) that has been declared but not assigned.