vx:  Add safe eval() function
This is eval().  Unlike other languages, V provides eval() access to the
keyword namespace only -- library subroutines (including eval() itself),
global state, and even the local state of the routine calling eval() are
all off-limits.  The argument must be a single string with no NUL bytes.
The argument is evaluated in a new interpreter context and the result is
returned by eval().  If an error occurs, eval() will throw an exception,
except for memory exhaustion which is currently a fatal error.