vlib:  Allow assignment to lists of symbols
This implementation is quite liberal with regard to what it allows.  For
example, in addition to `var x, var y = 3, 60` (which assigns 3 to x and
60 to y), it also accepts `var x, var y = 3, 60, 900` (which assigns the
list (60, 900) to y) and `var x, var y = 60, 900; x, y += 30, 90` (which
adds 30 and 90 to x and y respectively, for final values of 90 and 990).

(Both of these should be considered experimental and may be removed with
no warning.)

What it doesn't permit is type annotations in the actual list assignment
(though an easy workaround is to separate the declaration and assignment
into different statements).