vx:  Add system() function and -Z option
Add unrestricted mode, enabled by the -Z switch (for "zero limitations",
"zero restrictions", and/or "zero ****s given"[1]).  In restricted mode,
system() can only be called with an empty argument list, whose result is
the empty list.  In unrestricted mode, system() can be passed either the
empty list or a C string (i.e. a string with no NUL bytes), and it calls
the native system() with either NULL or the string pointer respectively.
In the former case, it returns either true or false; in the latter case,
it returns zero on success and throws an exception otherwise.  A failure
could be any of a non-zero exit status, a fatal signal, or an error from
trying to fork() in the first place.

[1] "cares", "hoots", etc.