vc:  Avoid cascade of recursive destruction
Pairs are mostly used like Lisp `cons` cells, with one atom on the left,
and a chain of pairs on the right -- so in practice, it's a linked list.
The default strategy for destruction recursively destroys members first,
causing a cascade of recursive destructor calls as long as one member is
a pair with a refcount of one.  With a million-atom list, this blows the
stack.  Destroy the list iteratively instead.