vc:  Replace function IDs with function pointers
Move the built-in functions to functions.cc.  Tweak function_from_name()
to return a function_type (a function pointer) instead of a function_id,
which is then used to construct a value in Parser::receive_token() which
will be of the same form as 'library' functions.  Remove function_id and
calc_function(), which are no longer used.