vlib:  Use vbox for Value type, int, string, proc
Use the new vbox storage type for Value's string and integer members, as
well as the proc (function object pointer) and the type field.  (The one
Value data member not included yet is the expression box, which requires
extent destructors.)

The vbox type contains an area for small-object-optimized storage and an
extent pointer for allocated objects, as well as a vbox type field which
determines whether an extent is allocated, and if not, what layout is in
use for the small object.  Additionally, a semantics field holds higher-
level type information that doesn't affect the storage.

The vbox class is layout-compatible with plus::string and plus::integer.
and conversion between each respectively to a vbox and back is supported
(if crudely).  The benefits of combining these into one type are triple:
(a) The vbox structure is smaller than the two collectively, since space
is not wasted on one when the other is in use.  (b) The resulting binary
is smaller since Value has to copy and destroy just the vbox, instead of
both the string and the integer (*at least* one of which is empty).  (c)
The interpreter runs measurably faster -- between 5% to 6% so, according
to a trial run in xv68k.  Moving the expression box into vbox will yield
even better performance.