vlib:  Place basic functions in the symbol table
Remove the plumbing for basic functions.  This is a double-edged axe (so
to speak):  On the plus side, the functions are now first-class objects,
so we can alias, bind, and compose them.  The caveat is that an argument
may no longer begin with a unary operator (unless in parentheses), since
it will be parsed as a binary operator instead, with the function as the
left operand since that's a value.  As an additional consideration, it's
foolish to maintain two unique syntactic behaviors for a single semantic
concept, and avoiding that is worth the cost of requiring parentheses in
`str (-x)` et cetera.