plus:  Move ibox_structure to header, reuse
This change removes the duplication between ibox and ibox_structure.  In
addition, it makes possible the future reuse of ibox_structure when vbox
is added to vlib.

It's unconventional to name a member "its".  This was chosen to minimize
the visual impact of changing every point of use of a data member.