plus:  Add ibox::move()
This is not actually an implementation of move semantics.  The one valid
use of ibox::move() obtaining a reference to a POD type for constructing
a vbox instance with an integer value.

This change unfortunately grants unsafe access to an ibox structure, but
so does a flagrant reinterpret-cast.  As always, callers are responsible
for protecting their feet from their own firearms.