vlib:  Reverse comma associativity
This doesn't affect the order in which list items are evaluated (left to
right), or the resulting list structure (since lists are normalized into
a canonical (first, rest) form).  What it does do is to defer processing
the list until a lower-precedence operator is seen -- which dramatically
improves parsing performance even for lists of fewer than a dozen items.