vlib:  Add lexical scopes and garbage collection
Localize symbols to the block scope in which they're declared.  Allow an
existing declaration in an outer scope to be shadowed by one in an inner
scope.  (Keyword-nature symbols live in their own symbol table and can't
be overridden.)

While parsing, record newly declared symbols in the symbol table for the
active lexical scope, and use that scope as well as its parent scopes to
resolve symbol names.  Instead of inserting a direct symbol value at the
point of use, substitute a symbol descriptor, consisting of the depth of
the lexical scope (relative to the active one) in which the found symbol
resides, and its index within that scope (in order of declaration), with
`_` always at depth 0, index 0.  Return both the block's expression tree
and the list of symbols it declares.

To begin execution, bind the root block's symbols to an empty stack, and
use the result as the initial stack.  Bind the stack to invocable values
(including both while-do condition expressions and block scopes).  Build
a new stack frame for block scopes using the scope's symbol presets, but
replace the `_` placeholder in the presets with the value in the current
stack frame.  (This will only be used if the block is *entered*, without
passing arguments as a *call* does.)

When invoking a block (either by entering or calling it), unshare all of
the symbol presets -- the block object includes a list of each symbol it
declares, and to make these actually be local, a fresh copy for each use
of the block is required.  For calls, use the argument list as the value
of `_` in the new stack frame.  For non-call block entry, reuse `_` from
the stack, without unsharing it.

Track each symbol as it's assigned (including copies made by unsharing),
to allow subsequent garbage collection of cyclically referenced objects.
(For now, garbage collection doesn't occur until the interpreter exits.)