vlib:  Support right-cantilevered binary operators
The subscript operator is a binary operator whose syntax consists of two
separate tokens straddling the right operand.  To contend with scenarios
such as this, recognize a leading substring composed of bytes with their
high bits set (if any) as the outer token (minus the high bits), and the
remainder as the inner token.