plus:  Silence GCC warnings in pseudo-dead code
Old versions of GCC complain about comparisons that are always false (or
true) due to the limited range of a type, in template code parameterized
*on that very type* -- so it's not *always* the case, since the compiler
is only considering one particular instantiation at a time.  Many of the
warnings occur in segments of code that aren't even used (i.e. dead) for
the instantiation in question.

Appease GCC by confusing it with multiplications by boolean values.  The
values are always true in the event the code is actually used (therefore
having no effect) and sometimes false otherwise (resulting in comparison
against zero instead of some large value, which suppresses the warning).