vc:  Axe the PC 'register'
Having while loops makes the PC 'register' variable redundant.  It won't
be supported in vx.  Using it correctly is not intuitive, and it assumes
a global symbol table shared between blocks which is about to be limited
to language keywords and constants.

Also, since the loop is now bounded, we don't need periodic_yield(), and
since each block is executed at most once, there's no reason to save the
parse trees.