vlib:  Add symbol descriptor values
A symbol descriptor (symdesc) is a tuple (represented as two 16-bit ints
packed into a 32-bit field) indicating stack depth (zero for the current
frame, and increasingly positive numbers for deeper frames) and an index
into the symbols local to the selected frame.  When encountering symbols
during parsing, each name is resolved against the linked list of lexical
scopes in effect into a symbol descriptor.  When the code block is later
invoked, symbol descriptors are resolved against the actual call stack.