vlib:  Report non-ASCII chars in sequences
We could have just written `if ( (unsigned char) *p < ' ' )` and avoided
the issue of reporting non-ASCII code points as control characters.  But
since non-ASCII bytes are likely to be UTF-8-encoded Unicode data, treat
a consecutive sequence of them as one token to prevent display artifacts
when printing the diagnostic.