vlib:  Add `lambda` and `return` operators
The lambda operator takes a block and returns a lambda, which is another
function-like object.  The difference is that lambdas are more function-
like, in that `break` and `continue` transfers reaching a lambda context
yield an error instead of passing through.  Additionally, a new `return`
transfer (initated by the new `return` operator) is also caught, and its
value is used as the lambda's result, whereas it jumps out of blocks the
same way `break` and `continue` do.