quickdraw:  Fix rect/region intersection clipping
Given a rectangle whose bottom intersects a rectangular subregion of the
region supplied, we were clipping the intersection to the bottom of that
rectangular region, not the input rectangle.

For example, for the one-pixel horizontal line at the bottom of the menu
bar and the gray region (i.e. the screen minus the rounded corners), the
intersection should be simply the unmodified line (with a bottom of 20),
but instead the bottom was 337 -- the region's next V-coordinate.