Pedestal:  Add ClickTarget (like CurrentFocus)
The click target, if any, is the view which received the last mouseDown.
We can't just dispatch mouseUp events by hit-testing (the way we do with
mouseDown events), because a view that tracks a click needs to know when
(and where) the mouseUp occurs, regardless of which view lies underneath
when that happens.  Conversely, the view in which the mouseUp occurs has
no need whatsoever to know about it, unless it was the one receiving the
corresponding mouseDown.