TestApp:  Make generally buildable with A-line
Add an A-line.conf and conditionalize for Carbon and Mach-O development.
The result of building for 68K using Metrowerks tools isn't yet runnable
in xv68k, since the startup code calls traps we haven't implemented yet.
But this change does make it easier to verify TestApp itself -- to check
that it behaves as expected in Mac OS.