macos:  Block only slightly in _GetNextEvent
Previously, we were preventing an application that called GetNextEvent()
in a loop from consuming all available CPU, by not returning null events
except when the mouse moved.  While resource-conservative, this is going
to be a problem for applications which need to operate in the absence of
user input.  A less drastic countermeasure is to throttle GetNextEvent()
to wait a short time before returning a null event.