[SEC] v68k-syscalls:  Fix memory access byte count
If n is a number of bits, then (n + k) >> 5 is a number of 32-bit units,
so multiply by 4 to get a byte count.

This had no effect on correctly written programs, though a malicious one
could have abused it by passing an fd count up to four times as large as
the number of bits available in the provided fd_set buffers, overwriting
whatever lay beyond in native memory space.