macos:  Add desktop-only _ClipAbove and _PaintOne
Add minimal _ClipAbove and _PaintOne trap implementations that work with
a NULL window argument (indicating the desktop).  Explicitly clip to the
bezel region to draw in the menu bar.  (This is a departure from actual,
observed Mac OS behavior where DrawMenuBar() fills in the rounded screen
corners -- ours doesn't, relying on clipping to the bezel region to keep
them black.)