d68k:  Fix CMPA decode
We were conflating the two-bit value extracted from the CMPA opcode with
the index into a pair of three-element arrays.  When the value is three,
the index is pulled from elsewhere (which we did correctly), but we made
the error of subsequently comparing the *index* to three, instead of the
original two-bit value, to determine which instruction to print.