[NEW] macrsrc:  Resource Manager emulation module
This is the start of an implementation of the Resource Manager.  It runs
within a 68K emulator and relies on the Freemount file transfer protocol
to retrieve data from the outside world.  It reads file descriptor 6 and
writes file descriptor 7, both of which must be connected to a Freemount
server.  Running `tcpclient $host $port xv68k ...` is one way to achieve
this.

For a resource of type 'ABCD', id=1234, GetResource() will construct the
path "System/r/04D2.ABCD" -- which is the scheme used by MacRelix to map
resources into directories and files -- and send a file-get request.  If
an exception is thrown, it sets ResErr to resNotFound (or memFullErr, if
appropriate) and returns NULL.  Otherwise, it returns a new handle after
copying the data.

Running freemountd in MacRelix will conveniently just work, in the sense
that requests will be answered using resources in a file "System" at the
server's top level.  Another option is to use a directory tree of actual
files, since macrsrc doesn't make a distinction.  A Mac-OS-hosted server
that calls its own native Resource Manager is also possible.

NOTE:  A future version of Freemount may directly support Mac resources,
and in any event, the path representation used by macrsrc (and MacRelix)
is subject to change.