splode:  Remove pointer_event_buffer's button slot
This reduces the payload length (and also the message type) from 4 to 3.

We were intending to use message type 3 for key events, but that doesn't
leave room for a device field, which we'll almost certainly want to have
later (since one byte is occupied by the character itself).  The pointer
event message, on the other hand, can get by with a merged device/button
field -- four bits each gives us 15 devices, each with 15 buttons (which
should be enough for anybody).  Also, the field's semantics haven't been
nailed down, so a simple one-button or three-button mouse might only use
one or two bits (respectively) using a Huffman coding scheme.  To cite a
contrived example, a game whose goal is to perform a 21-gun salute would
need 21 unique pointing devices, but each would have just a trigger (and
possibly a safety switch).