macos:  Add _LoadSeg trap, fix _Launch
Instead of loading and jumping into code segment 1 in _Launch, jump into
the jump table's first entry, which invokes _LoadSeg, loads the segment,
and jumps to the address of the start routine in the now-loaded segment.

Multi-segment applications are now supported, as well as apps whose main
entry point doesn't begin segment 1.