TestApp:  Remove build.pl script for SC compiler
We were only using SC in the first place because it allowed us to create
a CODE resource with no boilerplate start code.  Now that macrsrc exists
to provide GetResource(), Metrowerks' start code isn't a problem.  While
it would be decent to support more than one 68K compiler, SC is annoying
to work with (particularly in its handling of global data) and not worth
the effort, in my opinion.