TestApp:  Fix the run script
This script has a couple of problems:  One, it tries to execute the CODE
resource directly instead of invoking the Segment Loader.  Two, it isn't
connecting to a Freemount server to fulfill the GetResource() calls made
by the Metrowerks start code.

Use the graft tool to launch xv68k as a client connected to a freemountd
instance.  Use the app tool as the program to run in xv68k.

In order for this to work, you'll need to have a "68k" directory in your
home directory (which is already assumed, since that's where modules are
searched for).  In addition to the macos and macrsrc modules, there must
also be the app tool and the TestApp application, built for 68K.