xv68k:  Remove the unsynced alias of screen memory
This was an attempt to accelerate QuickDraw graphics by greatly reducing
the number of msync operations, from one per write to one per primitive.
This worked well enough for things like painting the entire desktop, but
it wasn't composable in the sense of syncing multiple related primitives
(as for drawing a window frame) all at once, unlike the screen lock that
replaces it.

This reverts commit bc6c34afd0d63e1b9ff295f540841afaa369f245.