macos:  Upgrade screen locks to raster locks
A raster lock combines a screen lock and a cursor lock.  When active, it
suppresses screen updates and also hides the cursor, the latter of which
is necessary for any drawing operations meant to ignore the cursor, i.e.
all of them other than screenshots or cursor drawing itself.

Screen locks that contains nested cursor locks (usually as raster locks)
don't need to be upgraded to raster locks, strictly speaking, but it may
be desirable to do so anyway, for performance.  There won't be a visible
difference, but using a raster lock at high level prevents the low-level
locks from hiding and showing the cursor repeatedly (since it'll already
be hidden), which won't be visible but will needlessly consume CPU.