xv68k:  Only call msync() in MacRelix
On Unix systems (or at least on Linux and OS X), msync() isn't necessary
to propagate writes in an mmapped memory region in one process to a view
of the underlying file in another process -- writes are immediately seen
by a subsequent pread(), for example.  On the other hand, msync() *does*
cause synchronization to physical media, which at best is not useful and
at worst is harmful to the media (e.g. flash storage), and in any event,
reduces performance to some degree.

MacRelix, however, relies on msync() to synchronize a bitmap view's bits
file with the screen.