xv68k:  Add --raster synchronization option
If --raster is used instead of --screen, append a metadata footer to the
screen file which includes a "sync relay" -- a structure that contains a
process-shared mutex and condition variable.  Since the mutex is useless
for synchronization until it's initialized, use a FIFO for synchronizing
initial access to the mutex.

I'm not entirely happy with the quality of this code, but I've spent too
long looking at it to effectively spot further issues for now.  I expect
to factor out the boilerplate raster creation code at some future point.