[NEW] kdgraphics, kdtext:  KDSETMODE wrappers
These replace pausecursor, which was designed from a misunderstanding of
what KDSETMODE does.  To use tty1 for graphics, run kdgraphics to switch
modes, and leave it that way.  If you intend to keep tty1 as a terminal,
then allocate a separate virtual console for graphics, e.g. tty7 -- just
like X11 and mac-on-linux.

Unlike pausecursor, kdgraphics and kdtext open tty0 and therefore affect
the current console, not necessarily tty1.

These programs will work for unprivileged users if they are setuid root.