vlib:  Let a dot disambiguate bare function calls
In V, you can't write `print -1` because it parses as binary subtraction
instead of a function call.  As a convenience, allow `print .-1` to work
instead.  Since `.` is generally followed by an initial (either a letter
or an underscore), changing the meaning of a dot followed by punctuation
doesn't affect supported usage.  (This idiom is whitespace-sensitive, so
`print . -1` will fail as before.)