vlib:  Add `data` type
A data value is much like a string, and even has identical storage.  The
difference is that data values are not expected to be printable text, so
they automatically print in hexidecimal.  For example, data("Nu") prints
`4e75`, and its representation is `x"4e75"`.  To convert a data value to
a string verbatim, use the data::string member.