vlib:  Replace "juxtaposition" with "adjacency"
"Juxtaposition" means to place things together to invoke contrast.  This
is an inappropriate term, since (a) following a function by its argument
list (for example) is strictly functional (so to speak), and has nothing
to do with demonstrating a contrast between them, nor any other literary
purpose; (b) as this is merely a syntactic construct, there's no intent,
per se, of any kind to speak of; and (c) the notion of placement implies
motion over time, whereas V source code is entirely static as far as the
parser is concerned -- nothing gets placed anywhere because everything's
already in place on arrival.  Therefore, replace "juxtaposition", an act
of intentionally moving things, with "adjacency", a descriptive property
indicating close proximity.