vlib:  Add data::size member
It's "size", as in C++'s sizeof operator, not "length", as in Pascal and
Perl's length() and C's strlen().  The former is a number of bytes while
the latter are numbers of elements.  (Even in the Pascal and C cases, in
which bytes are counted, the length byte or terminating NUL is omitted.)
It's true that C++ has size() as an element count, which is unfortunate,
but it's the odd one out.