vlib:  Add null value as a special type
The null value is a base type with a few modifications:  For one, `null`
happens to be its own type (although `null isa type` remains true).  For
another, unlike any other base type, null converted to boolean is false.
The only value of null type is null itself.  In these respects, it bears
similarity to empty-list, except that `() isa type` is false.  The major
difference between () and null is arity -- () has zero items and null is
one item.  Therefore, null survives interpolation into lists -- which is
essential for interoperability with JSON and SQL.