vlib:  Evaluate `STRING foo` as `str(STRING, foo)`
This is the same operator as the implicit function call.  When the first
operand is a string, the operator performs string concatenation instead.
Since it associates from right to left, when chained, each operand other
than the last must be a string.  The string operands need not be literal
string tokens -- expressions evaluating to string values work just fine.

In practice, this is useful for one-liner scripts and being less verbose
generally.  Compare:

	print "foo: " foo;

	print("foo: ", foo);

	print ("foo: ", foo);

	print( "foo: ", foo );

Using the implicit operator here eliminates the need for parentheses (as
well as 40% of the Shift key use), making the code less cluttered.