vlib:  Fix double-free in pair_destructor()
Explicitly unset the extent destructor of secondary objects destroyed in
pair_destructor() before calling extent_release() to free their memory.

In pair_destructor(), which is the extent destructor for Expr structures
in linked lists, we destroy the lists iteratively instead of recursively
to prevent stack overflow.  Since the function calling pair_destructor()
deletes the memory afterward, we refrain from doing so, but we do delete
the memory of secondary Expr objects after destructing them, since we're
substituting for a use of the delete operator in that case.

The problem was that extent_release() calls the extent destructor, which
was set to pair_destructor().  Aside from not wanting recursion, this is
problematic because pair_destructor() calls the Expr destructor -- which
we just called in the outer pair_destructor() for the same object.

In practice, the vbox destructor shields itself from double destruction,
so it's not affected.  However, the source_spec contains a string object
for the pathname, which very much is affected -- if the pathname is long
enough to cause an allocation, then after executing the program, vx will
double-free the pathname and in all likelihood crash.