vlib:  Implement array slicing
A slice is a subscript operation whose left operand is an array (or some
variant thereof, like a string) and whose index is an array of integers.
The result is a new array, whose members are those selected by the index
array.  For any array `foo`, and any array `bar` whose members are valid
indices in foo, `foo[bar] == bar map {foo[_]}` is true.  More generally,
`foo[bar] == (typeof foo)(bar map {foo[_]})` is true even when foo isn't
strictly an array, but a string or data unit.