vlib:  Fix vbox::has_extent() always being false
On a Raspberry Pi with g++ 4.9.2, vbox::has_extent() never returns true,
even though the value it's comparing to zero is of type vbox_type, which
is an enum with a member whose value is negative.  Explicitly cast it to
int8_t as a workaround.  Also, delete vbox::type(), which is now unused.