vx:  Add Project Euler problem #7 example
The performance is disappointing, but I'm merging this anyway.  First of
all, this code is blameless (since it just calls nth_prime(), which does
all the work).  Second, it's not obvious which optimizations nth_prime()
requires to speed things up.  Accelerating the remainder operation helps
a little, as does using contiguous array storage, but it still clocks in
at over three seconds on this machine, while my Perl code uses under 100
milliseconds.