vlib:  Recognize `end.` token
When preceded and followed by newlines, `end.` marks the end of a source
file.  Anything that follows will be completely ignored and is therefore
free from V syntax requirements.

Note that it's still conceivable that a translator might expect a source
file to contain consistent newlines or well-formed UTF-8, and failure to
conform might abort processing prior to V tokenization.  It's the user's
responsibility to ensure the compatibility of any appended data with the
software that processes it.