etc/startup:  Use test-type for faster startup
This isn't strictly a no-op.  Previously we were checking for the x bit;
now we check for regular files.  Both are preconditions, and now instead
of erroneously proceeding when the file is a directory, we'll mistakenly
continue when it's a non-executable regular file.  However, the scenario
of no file at all, which is still detected, is the only really important
one -- the others are effectively edge cases.  The priority is improving
performance.