Genie:  Listen on a Unix-domain socket at startup
On OS X 10.4 at least, this fails as is because a global variable in the
Pedestal library doesn't get statically initialized before it gets used,
even though it's defined in the same translation unit -- a compiler bug,
in all likelihood.  If you ignore gNewWindow and just call NewCWindow(),
it works, but then the application segfaults on quit.  (This is a static
destruction ordering issue that should go away after FORGE is revised to
individually allocate view properties.)  It's all good in 10.8, however.