Pedestal, apps:  Factor out close_window()
When an application wants to close a window, close_window() is the right
function to call.  It currently calls Window's Close(), which then calls
the window's close handler (which is responsible for deleting the window
owner, which also releases the window).  (It should only be called after
determining that the window should, in fact, be closed.)