Pedestal:  Add WindowClosed_proc
A 'closed' proc takes precedence over a close handler in close_window().
If the former is set, the latter will not be called.  Therefore, it must
ensure that the window is destroyed, as well as any auxiliary resources,
if any.  (Pre-Carbon code requires custom window storage for this, but a
Carbon-only window's 'closed' proc could simply call DisposeWindow(); if
it's Mach-O-only, it could even pass DisposeWindow() directly.  However,
this will be the default in the future anyway.)