UseEdit:  Factor out id_of_window()
We no longer reuse the document token's AEDesc handle and just splice in
a new descriptorType.  Now we create a new AEDesc for the keyData.  This
is not only a nicer way to do things (if less efficient), but it sets us
up for having a non-identity address-to-ID conversion function.