UseEdit:  Shift window IDs right by three bits
Don't use memory addresses verbatim.  AppleScript only supports integers
up to 29 bits, above which it substitutes floats automatically.  Instead
of trying to make the scripting subsystem accept a floating point object
ID, reduce the range of the conversion function so th ID will always fit
in 29 bits.  This works as long as the input addresses are distant by at
least eight bytes.  Pre-Carbon, this was certain to be the case, since a
WindowRef points to a WindowRecord, which is a lot larger than that.  In
Carbon there are no guarantees, though it doesn't seem to be an issue so
far.