Pedestal:  Install a WindowClose handler in Carbon
This is necessary to make the Close command appear in the popup menu for
minimized windows in the Dock.  It will also be needed to allow a window
in compositing mode to be closed by clicking the close box even when the
application isn't frontmost.  (A mouse-down event is generated if either
the window is non-compositing or the application is frontmost (or both),
but not otherwise.)