Genie:  Start the gui service when entering main()
This avoids a crash on Mac OS X 10.4 when opening the console log window
(e.g. because ~/var/run/fs doesn't exist) due to an uninitialized static
variable (which seems to be a compiler bug).

Note that although replacing the use of gNewWindow with a simple call to
NewCWindow() or CreateNewWindow() sidesteps the crash on launch, there's
still a crash on exit lurking due to order-of-destruction issues.