vlib:  Rename `data` type to `vector`, alias `vec`
The identifier 'data' is too useful as a variable name to be claimed for
a built-in type.  Even though it's too vague and amorphous to be used in
most cases, occasionally it's a perfect fit.  Instead, use `vector` (and
`vec`).  It's the same length as `string`, and the parallel to C++ types
will make it easier to understand.